requirements.txt
asgiref==3.7.2
Django==5.0.2
django-database-url==1.0.3
django-jazzmin==2.6.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
